Move v1 math utils to v1 file
