+++ src/pages/v1/create/ConfirmDeployProject.tsx
@@ -24,7 +24,6 @@ import {
   perbicentToPercent,
   permilleToPercent,
 } from 'utils/formatNumber'
-import { amountSubFee } from 'utils/math'
 import { orEmpty } from 'utils/orEmpty'
 import {
   getUnsafeV1FundingCycleProperties,
@@ -32,6 +31,7 @@ import {
   hasFundingTarget,
   isRecurring,
 } from 'utils/v1/fundingCycle'
+import { amountSubFee } from 'utils/v1/math'
 
 import { V1CurrencyName } from 'utils/v1/currency'
